requirements.txt
# Pinned 2026-04-03 — PyPI latest stable at lock time.
# robotframework-seleniumlibrary pulls selenium; explicit pin keeps CI/env aligned.

# --- Robot + browser stack ---
robotframework==7.4.2
selenium==4.41.0
robotframework-seleniumlibrary==6.8.0

# --- Other Robot ecosystem ---
robotframework-pabot==5.2.2
robotframework-tidy==4.18.0
robotframework-faker==6.0.0
robotframework-requests==0.9.7
allure-robotframework==2.15.3

# --- Build / tooling ---
setuptools==82.0.1

# --- Salesforce API data seeding ---
simple-salesforce==1.12.6

# --- App & AI ---
python-dotenv==1.2.2
streamlit==1.56.0
openai==2.30.0
google-generativeai==0.8.6
